Fix header background bug on mobile and move it to separate components
