+++ README.md
@@ -5,6 +5,7 @@ This is my personal site – A place for me to showcase my work, explore ideas a
 ## Todo
 
 - [x] Complete initial style guide and design for Start page
+- [x] Fix header background bug on mobile and move it to separate components
 - [ ] Implement sticky footer
 - [ ] Add OG-tags and NEXT SEO
 - [ ] Create "Now playing" component from Spotify API
